refactor(workstation): the cross-window observation surface gets its own boundary (#18460)
AC-1's Contract Ledger row for receipts and diagnostics asks for truthful
observation at the owning boundary. `readCrossWindowGestureSnapshot` had no
boundary: 70 lines of read-only projection sitting in a 3,042-line production
class with zero engine readers, consumed only by the tour driver and five test
files.

It could not become a helper module. Four e2e specs invoke it across the Neural
Link BY NAME at seven call sites —
callMethod(workspaceId, 'readCrossWindowGestureSnapshot', [...]) — and a unit
spec reaches it through Workspace.prototype. Both resolve against the prototype,
so it has to stay an instance method; a module plus a one-line delegate would
satisfy the line count and reintroduce exactly the forwarding facade AC-2
forbids. A mixin keeps the instance contract and still moves the code out.
Neo.dashboard.dock.window.TopologySeams is the same shape for the same stated
reason: a seam its caller resolves by name has to be an instance method.

WHAT DID NOT MOVE, and this corrects my own AC-1 framing. I called the four
vessel receipts "an observation surface, not product state" and asked whether
they belonged in production source at all. Reading them says otherwise: each is
a stage machine its owning method amends as it advances -- identity-refused,
route-refused, embodiment-refused, native-dispatched, acknowledged, with threw
branches -- and closeTearOutVessel holds its receipt in a local const precisely
so those amendments survive re-entrancy. That is the method's own progress
record written across awaits, and relocating it would be the "lost success or
recovery claim after an await" the same Ledger row refuses. Only the read
projection moved.

No deletion is claimed. I measured which snapshot fields consumers read and the
instrument was a substring grep over whole spec files, which counts any
occurrence of a token rather than a read of that field -- a proxy, and not one
that can justify removing payload. Declining the deletion explicitly rather
than inventing one.

That same proxy then leaked into this message's own reader counts. The first
version said "eight test files" and "five e2e specs"; enumerated by call site
the truth is five test files, four of them e2e. The bad instrument was named
one paragraph above the numbers it produced, which is the useful part of the
finding: disclosing an instrument's weakness does not quarantine its output.

One transcription defect caught before it shipped: moving 70 dense lines by hand
put `indicatorSet.activePreviewId` where the original reads
`indicatorMenu.activeCandidate?.preview?.previewId`. That would have left
snapshot.ready comparing against undefined forever, so every film executor's
pre-release gate would hang. A line-by-line diff of the moved body against the
original found it; nothing else would have.

apps/workstation/view/Workspace.mjs: 3,042 -> 2,963 physical, 1,689 -> 1,624
code. Baseline ratcheted to match. New mixin is 134 lines including its own
rationale.

Evidence: workstation unit tier 74 passed at this head, including the spec that
reaches the method through Workspace.prototype. WorkstationNativeTitlebarDragNL
green with NEO_AGENTOS_RUNTIME_ROOT set -- that arm is one of the four that call
it by name over the Neural Link. Without the runtime root that config selects
ZERO specs and reports success, so the green is stated with the variable named.

The app had no gitignore whitelist: /apps/**/*.mjs ignores everything and
every tracked workstation file predates the rule, so git never applied it to
them. A NEW file there is silently unstageable, which is how this surfaced --
`git add` refused it. Three negation lines land in this commit, matching the
shape every other app already has, and they expose exactly one previously
ignored path: the file this commit adds.
