+++ apps/workstation/view/Workspace.mjs
@@ -1,28 +1,29 @@
-import Component                from '../../../src/component/Base.mjs';
-import Container                from '../../../src/container/Base.mjs';
-import DockWorkspace            from '../../../src/dashboard/dock/Workspace.mjs';
-import PopupWorkspace           from './PopupWorkspace.mjs';
-import Feed                     from '../store/Feed.mjs';
-import FeedPane                 from './FeedPane.mjs';
-import Scale                    from '../store/Scale.mjs';
-import ScalePane                from './ScalePane.mjs';
-import DockDragAffordances      from '../../../src/dashboard/dock/interaction/DragAffordances.mjs';
-import DockDropIndicators       from '../../../src/dashboard/dock/interaction/DropIndicators.mjs';
-import DockLayoutAdapter        from '../../../src/dashboard/dock/projection/LayoutAdapter.mjs';
-import PerspectiveLibrary       from '../../../src/dashboard/dock/persistence/PerspectiveLibrary.mjs';
-import TopologyLibrary          from '../../../src/dashboard/dock/persistence/TopologyLibrary.mjs';
-import Placement                from '../../../src/dashboard/dock/window/Placement.mjs';
-import DockPreview              from '../../../src/dashboard/dock/interaction/Preview.mjs';
-import DockProjectionReconciler from '../../../src/dashboard/dock/projection/Reconciler.mjs';
-import DockService              from '../../../src/ai/client/DockService.mjs';
-import WorkspaceDocument        from '../../../src/dashboard/dock/model/WorkspaceDocument.mjs';
-import WorkspaceController      from './WorkspaceController.mjs';
-import Operations               from '../../../src/dashboard/dock/model/Operations.mjs';
-import Persistence              from '../../../src/dashboard/dock/model/Persistence.mjs';
-import StateProvider            from '../../../src/state/Provider.mjs';
-import TourToolbar              from './TourToolbar.mjs';
-import TransactionManager       from '../../../src/manager/Transaction.mjs';
-import WindowManager            from '../../../src/manager/Window.mjs';
+import Component                  from '../../../src/component/Base.mjs';
+import Container                  from '../../../src/container/Base.mjs';
+import DockWorkspace              from '../../../src/dashboard/dock/Workspace.mjs';
+import PopupWorkspace             from './PopupWorkspace.mjs';
+import Feed                       from '../store/Feed.mjs';
+import FeedPane                   from './FeedPane.mjs';
+import Scale                      from '../store/Scale.mjs';
+import ScalePane                  from './ScalePane.mjs';
+import DockDragAffordances        from '../../../src/dashboard/dock/interaction/DragAffordances.mjs';
+import DockDropIndicators         from '../../../src/dashboard/dock/interaction/DropIndicators.mjs';
+import DockLayoutAdapter          from '../../../src/dashboard/dock/projection/LayoutAdapter.mjs';
+import PerspectiveLibrary         from '../../../src/dashboard/dock/persistence/PerspectiveLibrary.mjs';
+import TopologyLibrary            from '../../../src/dashboard/dock/persistence/TopologyLibrary.mjs';
+import Placement                  from '../../../src/dashboard/dock/window/Placement.mjs';
+import DockPreview                from '../../../src/dashboard/dock/interaction/Preview.mjs';
+import CrossWindowGestureSnapshot from './CrossWindowGestureSnapshot.mjs';
+import DockProjectionReconciler   from '../../../src/dashboard/dock/projection/Reconciler.mjs';
+import DockService                from '../../../src/ai/client/DockService.mjs';
+import WorkspaceDocument          from '../../../src/dashboard/dock/model/WorkspaceDocument.mjs';
+import WorkspaceController        from './WorkspaceController.mjs';
+import Operations                 from '../../../src/dashboard/dock/model/Operations.mjs';
+import Persistence                from '../../../src/dashboard/dock/model/Persistence.mjs';
+import StateProvider              from '../../../src/state/Provider.mjs';
+import TourToolbar                from './TourToolbar.mjs';
+import TransactionManager         from '../../../src/manager/Transaction.mjs';
+import WindowManager              from '../../../src/manager/Window.mjs';
 import {
     createDockVesselEmbodiment,
     createDockVesselProxyEmbodiment
@@ -139,6 +140,15 @@ class Workspace extends DockWorkspace {
          * @protected
          */
         className: 'Workstation.view.Workspace',
+        /**
+         * The cross-window observation surface. A mixin rather than a helper module because every
+         * reader addresses it by name on the instance — the Neural Link resolves
+         * `readCrossWindowGestureSnapshot` on this component, and a unit spec reaches it through
+         * `Workspace.prototype`. Same shape and same reason as
+         * {@link Neo.dashboard.dock.window.TopologySeams}.
+         * @member {Neo.core.Base[]} mixins=[CrossWindowGestureSnapshot]
+         */
+        mixins: [CrossWindowGestureSnapshot],
         /** @member {Neo.controller.Component} controller */
         controller: WorkspaceController,
         /**
@@ -2386,95 +2396,6 @@ class Workspace extends DockWorkspace {
         me.vesselParkHandlers.onVesselRetired({itemId, retirement: true});
         me.nativeVesselParkHandlers.onVesselRetired({itemId, retirement: true});
         recovered && me.retireVesselWorkspaceTarget(itemId)
-    }
-
-    /**
-     * @summary Reads the semantic, rendered, arbitration, and physical-park truth for one
-     * cross-window pointer frame.
-     *
-     * The film executors use this as their pre-release gate: mouseup is withheld until the
-     * coordinator has exactly one stable claim, its winning target is engaged, and the target's
-     * semantic preview equals the preview rendered in that window. A converting tear-out can add
-     * `parkedItemId`, which additionally requires the exact source vessel to be strictly parked.
-     * Visuals are read from the current participant's supplied or already-created controller;
-     * observing readiness never creates the popup's lazy preview tier.
-     * @param {Object} context
-     * @param {String|null} [context.parkedItemId=null]
-     * @param {Neo.dashboard.dock.interaction.TabSortZone|null} [context.sourceZone=null]
-     * @param {String|null} [context.sourceZoneId=null] Clone-safe Neural Link alternative.
-     * @param {String} context.targetWorkspaceId
-     * @returns {Object}
-     * @protected
-     */
-    readCrossWindowGestureSnapshot({parkedItemId=null, sourceZone=null, sourceZoneId=null, targetWorkspaceId}={}) {
-        sourceZone ??= sourceZoneId ? Neo.get(sourceZoneId) : null;
-
-        let me            = this,
-            isMain        = targetWorkspaceId === Workspace.MAIN_WORKSPACE_ID,
-            state         = isMain ? null : me.getPopupState(targetWorkspaceId),
-            participation = isMain ? me.crossWindowParticipations.get(targetWorkspaceId) : state?.host?.participation,
-            affordances   = participation?.affordances ?? participation?.ownedAffordances,
-            target        = participation?.target,
-            coordinator   = sourceZone?.dragCoordinator,
-            arbiter       = coordinator?.pointerClaimArbiter,
-            winner        = arbiter?.resolve?.() ?? null,
-            semantic      = target?.currentPreview ?? null,
-            rendered      = affordances?.preview?.dockPreview ?? null,
-            indicatorMenu = affordances?.indicators,
-            indicatorSet  = indicatorMenu?.candidateSet ?? null,
-            sensor        = sourceZone?.vesselConversionSensor,
-            parkedVessel  = me.vesselParkHandlers?.parkedVessel ?? null,
-            sourceVessel  = parkedItemId && me.resolveTearOutVessel(parkedItemId),
-            targetProxy   = parkedItemId && me.vesselProxyEmbodiment.snapshot(parkedItemId),
-            snapshot      = {
-                claimCount: arbiter?.claimCount ?? 0,
-                converted : sensor?.converted === true && sensor?.transitioning !== true,
-                engaged   : coordinator?.activeTargetZone === target,
-                indicators: indicatorSet ? {
-                    activePreviewId: indicatorMenu.activeCandidate?.preview?.previewId ?? null,
-                    candidateCount : (indicatorSet.cross?.length ?? 0) + (indicatorSet.root?.chips?.length ?? 0),
-                    itemId         : indicatorSet.itemId ?? null,
-                    visible        : !indicatorMenu.cls?.includes?.('neo-dashboard-dock-drop-indicators-hidden')
-                } : null,
-                parkedItemId: parkedVessel?.itemId ?? null,
-                parkReceipt : me.lastVesselParkReceipt
-                    ? WorkspaceDocument.clone(me.lastVesselParkReceipt)
-                    : null,
-                preview              : semantic ? WorkspaceDocument.clone(semantic) : null,
-                rendered             : rendered ? WorkspaceDocument.clone(rendered) : null,
-                sourceVesselConnected: Boolean(
-                    sourceVessel?.windowId && Neo.manager?.Window?.get(sourceVessel.windowId)
-                ),
-                sourceVesselWindowId: sourceVessel?.windowId ?? null,
-                restoreReceipt      : me.lastVesselRestoreReceipt
-                    ? WorkspaceDocument.clone(me.lastVesselRestoreReceipt)
-                    : null,
-                targetProxy,
-                targetWorkspaceId,
-                winnerStableId      : winner?.stableId ?? null
-            };
-
-        snapshot.ready = snapshot.claimCount === 1
-            && snapshot.engaged
-            && snapshot.winnerStableId === targetWorkspaceId
-            && Boolean(snapshot.preview?.previewId)
-            && snapshot.rendered?.previewId === snapshot.preview.previewId
-            && snapshot.indicators?.activePreviewId === snapshot.preview.previewId
-            && (parkedItemId == null || (
-                snapshot.converted && snapshot.parkedItemId === parkedItemId &&
-                snapshot.sourceVesselConnected &&
-                snapshot.indicators?.itemId === parkedItemId &&
-                snapshot.indicators.candidateCount >= 5 &&
-                snapshot.indicators.visible &&
-                targetProxy?.itemId === parkedItemId &&
-                targetProxy.ownsPane &&
-                targetProxy.settled &&
-                targetProxy.sourceWindowId === snapshot.sourceVesselWindowId &&
-                targetProxy.targetWindowId === target?.windowId &&
-                targetProxy.visible
-            ));
-
-        return snapshot
     }
 
     /** @summary Tears down the workspace's view tree and owned docking resources. @param {...*} args */
