+++ buildScripts/util/file-size-baseline.json
@@ -1,5 +1,5 @@
 {
-    "apps/workstation/view/Workspace.mjs": 1388,
+    "apps/workstation/view/Workspace.mjs": 1323,
     "examples/dashboard/crossWindow/DemoBWorkspace.mjs": 2615,
     "src/dashboard/dock/Workspace.mjs": 1100,
     "src/main/addon/DragDrop.mjs": 1267
